+++ temba/utils/views.py
@@ -1,18 +1,11 @@
 import logging
-from math import ceil
 
 from django import forms
-from django.core.paginator import Paginator
 from django.db import transaction
 from django.http import HttpResponse
-from django.utils.functional import cached_property
 from django.utils.translation import ugettext_lazy as _
 from django.views import View
 from django.views.decorators.csrf import csrf_exempt
-
-from temba.contacts.models import ContactGroupCount
-from temba.utils.es import ModelESSearch
-from temba.utils.models import ProxyQuerySet, mapEStoDB
 
 logger = logging.getLogger(__name__)
 
@@ -150,89 +143,6 @@ class BaseActionForm(forms.Form):
 
         else:  # pragma: no cover
             return dict(error=_("Oops, so sorry. Something went wrong!"))
-
-
-class ContactListPaginator(Paginator):
-    """
-    Paginator that knows how to work with ES dsl Search objects
-    """
-
-    ES_SEARCH_BUFFER_SIZE = 10000
-
-    @cached_property
-    def num_pages(self):
-
-        if self.is_es_search():
-            # limit maximum number of pages when searching contacts on ES
-            # https://github.com/rapidpro/rapidpro/issues/876
-
-            record_count = self.count if self.count < self.ES_SEARCH_BUFFER_SIZE else self.ES_SEARCH_BUFFER_SIZE
-
-            return ceil(record_count / self.per_page)
-        else:
-            return super().num_pages
-
-    @cached_property
-    def count(self):
-        if self.is_es_search():
-            # execute search on the ElasticSearch to get the count
-            return self.object_list.count()
-        else:
-            # get the group count from the ContactGroupCount squashed model
-            group_instance = self.object_list._hints.get("instance")
-            if group_instance:
-                return ContactGroupCount.get_totals([group_instance]).get(group_instance)
-            else:
-                return 0
-
-    def _get_page(self, *args, **kwargs):
-        new_args = list(args)
-
-        es_search = args[0]
-
-        if self.is_es_search(es_search):
-            # we need to execute the ES search again, to get the actual page of records
-            new_object_list = args[0].execute()
-
-            new_args[0] = new_object_list
-
-        return super()._get_page(*new_args, **kwargs)
-
-    def page(self, number):
-        """Return a Page object for the given 1-based page number."""
-        number = self.validate_number(number)
-        bottom = (number - 1) * self.per_page
-        top = bottom + self.per_page
-
-        if top + self.orphans >= self.count:
-            top = self.count
-
-        # make sure to not request more than we can return, set the upper limit to the ES search buffer size
-        if top >= self.ES_SEARCH_BUFFER_SIZE and self.is_es_search():
-            top = self.ES_SEARCH_BUFFER_SIZE
-
-        return self._get_page(self.object_list[bottom:top], number, self)
-
-    def is_es_search(self, obj=None):
-        if obj is None:
-            return isinstance(self.object_list, ModelESSearch)
-        else:
-            return isinstance(obj, ModelESSearch)
-
-
-class ContactListPaginationMixin(object):
-    paginator_class = ContactListPaginator
-
-    def paginate_queryset(self, queryset, page_size):
-        paginator, page, new_queryset, is_paginated = super().paginate_queryset(queryset, page_size)
-
-        if isinstance(queryset, ModelESSearch):
-            model_queryset = ProxyQuerySet([obj for obj in mapEStoDB(self.model, new_queryset)])
-            return paginator, page, model_queryset, is_paginated
-
-        else:
-            model_queryset = ProxyQuerySet([obj for obj in new_queryset])
-            return paginator, page, model_queryset, is_paginated
 
 
 class ExternalURLHandler(View):